refactor: move get_config_option helper to a dedicated module in utils
